fix(deps): cero vulnerabilidades, y el repo pasa a tener UN solo formateador
`pip-audit` local pasa de 45 vulnerabilidades en 12 paquetes a NINGUNA.

Subidos: pytest 8.4.2->9.1.1, pytest-asyncio 0.26->1.4, pytest-cov 5->7.1,
paramiko 3.5.1->5.0.0 (tope `<4` levantado en apps/workers), black->26.3.1,
datamodel-code-generator, langsmith, pyjwt, pypdf, python-multipart,
starlette 1.0->1.3.1, setuptools->83, pip->26.1.2, OTel 0.48b0->0.65b0, ruff
0.5->0.16.3, y 52 pines sincronizados con lo instalado.

## Tres cosas que no eran una subida de versión

**`python-jose` estaba instalada y ya no se usa.** Era la única que arrastraba
`ecdsa`, cuya PYSEC-2026-1325 **no tiene versión que la corrija** — el único caso
que habría justificado una excepción en `.pip-audit-ignore`. Pero el código migró
a `joserfc` y sus propios comentarios lo dicen (`auth/jwt.py`: «python-jose is out
of the dependency set»); no está en ningún `pyproject` ni en `constraints.txt`.
Era residuo del venv. Se desinstala y la vulnerabilidad desaparece en vez de
excusarse.

**`setuptools 83` retiró `pkg_resources`** y tumbó 39 ficheros de test con
`ModuleNotFoundError` — no en nuestro código, sino dentro de
`opentelemetry-instrumentation 0.48b0`. Se sube OTel, que ya usa
`importlib.metadata`, en vez de bajar setuptools: el arreglo va donde está la causa.

**Se retira el hook de black.** Con black 26.3.1 (obligado por PYSEC-2026-2120 y
2121) y ruff 0.16.3, sus estilos dejaron de coincidir: black reformateaba 17
ficheros, ruff-format los devolvía, y así **sin converger jamás** — ningún `git
commit` podía terminar. No es el gotcha del comentario en la llamada encadenada:
son dos estilos distintos sobre el árbol entero, y ningún cambio en el código lo
arregla. Se queda `ruff-format` porque es el MISMO motor que el linter (una
versión, una config, un caché) y nació black-compatible; bajar black no era opción
porque su versión vieja es justo la vulnerable. El porqué queda escrito en el
propio `.pre-commit-config.yaml`, con la comprobación para quien quiera revertirlo.

## Las reglas de ruff que entraron solas

`select` nombra FAMILIAS (`PL`, `RUF`, `UP`), así que subir ruff matriculó 717
avisos que nadie eligió. Se decide una vez y en `pyproject.toml`, no con 717
`# noqa` que nadie vuelve a mirar. Rechazadas con motivo: `PLC0415` (669 casos —
imports locales deliberados, que rompen ciclos api_server.db ↔ routers y evitan
pagar el arranque de spacy/presidio/langgraph donde no se usan) y `PLR0917`,
hermana de la ya ignorada `PLR0913`. Adoptadas y saldadas: las 128 restantes.

Y una exención que NO es estilo: `N802` en `code_safety.py`. `ast.NodeVisitor`
despacha POR NOMBRE, así que el método tiene que ser `visit_Call`; en minúsculas
no se llamaría nunca y el analizador de seguridad dejaría de ver las llamadas
**en silencio** — un fallo mucho peor que el aviso.

**Un autofix de ruff dejó código ROTO**: `UP047` convirtió `run_remote_probe` a la
sintaxis PEP 695 declarando `[T]` y dejando las anotaciones en `_T`. Lo cazó mypy.

Verificación: pip-audit **sin vulnerabilidades** · unit+security 4580 passed ·
mypy 665 ficheros · ruff y ruff-format convergen en dos pasadas · pip check limpio.

+++ apps/api-server/src/api_server/auth/rate_limit.py
@@ -103,8 +103,8 @@ class RateLimiter:
         # The window frees a slot when the oldest in-window hit ages out.
         # Fall back to `now` if the set is somehow empty (defensive).
         oldest_score = oldest[0][1] if oldest else now
-        reset_at = int(math.ceil(oldest_score + window_seconds))
-        retry_after = max(1, int(math.ceil(oldest_score + window_seconds - now)))
+        reset_at = math.ceil(oldest_score + window_seconds)
+        retry_after = max(1, math.ceil(oldest_score + window_seconds - now))
 
         return RateLimitResult(
             allowed=allowed,
